use cleaner cp slot_mapping
